refactor(products): change route path
