+++ api/v1/routes/organization.py
@@ -5,7 +5,13 @@ from sqlalchemy.orm import Session
 
 from api.utils.success_response import success_response
 from api.v1.models.user import User
-from api.v1.schemas.organization import CreateUpdateOrganization, PaginatedOrgUsers, OrganizationBase
+from api.v1.schemas.organization import (
+    CreateUpdateOrganization,
+    PaginatedOrgUsers,
+    OrganizationBase,
+)
+from api.v1.services.product import product_service
+from api.v1.schemas.product import ProductCreate
 from api.db.database import get_db
 from api.v1.services.user import user_service
 from api.v1.services.organization import organization_service
@@ -19,7 +25,7 @@ organization = APIRouter(prefix="/organizations", tags=["Organizations"])
 )
 def create_organization(
     schema: CreateUpdateOrganization,
-    db: Session= Depends(get_db),
+    db: Session = Depends(get_db),
     current_user: User = Depends(user_service.get_current_user),
 ):
     """Endpoint to create a new organization"""
@@ -41,7 +47,7 @@ def create_organization(
 
 
 @organization.get(
-    '/{org_id}/users',
+    "/{org_id}/users",
     response_model=success_response,
     status_code=status.HTTP_200_OK,
 )
@@ -50,21 +56,19 @@ async def get_organization_users(
     db: Session = Depends(get_db),
     current_user: User = Depends(user_service.get_current_user),
     skip: int = 1,
-    limit: int = 10
+    limit: int = 10,
 ):
-    '''Endpoint to fetch all users in an organization'''
+    """Endpoint to fetch all users in an organization"""
 
-    return organization_service.paginate_users_in_organization(
-        db, org_id, skip, limit
-    )
+    return organization_service.paginate_users_in_organization(db, org_id, skip, limit)
 
 
-@organization.patch('/{org_id}', response_model=success_response, status_code=200)
+@organization.patch("/{org_id}", response_model=success_response, status_code=200)
 async def update_organization(
     org_id: str,
     schema: CreateUpdateOrganization,
     db: Session = Depends(get_db),
-    current_user: User = Depends(user_service.get_current_user)
+    current_user: User = Depends(user_service.get_current_user),
 ):
     """Endpoint to update organization"""
 
@@ -78,10 +82,31 @@ async def update_organization(
 
 
 @organization.get("", status_code=status.HTTP_200_OK)
-def get_all_organizations(super_admin: Annotated[User, Depends(user_service.get_current_super_admin)], db: Session = Depends(get_db)):
+def get_all_organizations(
+    super_admin: Annotated[User, Depends(user_service.get_current_super_admin)],
+    db: Session = Depends(get_db),
+):
     orgs = organization_service.fetch_all(db)
     return success_response(
         status_code=status.HTTP_200_OK,
         message="Retrived all organizations information Successfully",
-        data = jsonable_encoder(orgs)
+        data=jsonable_encoder(orgs),
+    )
+
+
+@organization.post("/{org_id}/products", status_code=status.HTTP_201_CREATED)
+def product_create(
+    org_id: str,
+    product: ProductCreate,
+    current_user: Annotated[User, Depends(user_service.get_current_user)],
+    db: Session = Depends(get_db),
+):
+    created_product = product_service.create(
+        db=db, schema=product, org_id=org_id, current_user=current_user
+    )
+
+    return success_response(
+        status_code=status.HTTP_201_CREATED,
+        message="Product created successfully",
+        data=jsonable_encoder(created_product),
     )
